Version 1.1.0-RC8 Version 4 Enabled the LCD info menu. Changed the move axis menu to select the axis first and distance second. Relocated the move axis menu item to the top of the prepare menu. Added filament change option to the prepare menu. Reduced the duration of the filament change beep from 300ms to 20ms. Added a quick menu with commonly used functions. Added code to prevent a z-bump if the axis hasn't been homed.  This is to prevent driving the build plate past Z Max. Added option to drop build plate from the prepare menu.

+++ Marlin/stepper.cpp
@@ -73,7 +73,7 @@ block_t* Stepper::current_block = NULL;  // A pointer to the block currently bei
 #endif
 
 #if HAS_MOTOR_CURRENT_PWM
-  int Stepper::motor_current_setting[3] = PWM_MOTOR_CURRENT;
+  unsigned long Stepper::motor_current_setting[3] = PWM_MOTOR_CURRENT;
 #endif
 
 // private:
